+++ scripts/new_version.py
@@ -28,28 +28,28 @@ def main(argv):
     new_version = ".".join(map(str, new_version_tuple))
 
     history_path = os.path.join(PROJECT_DIRECTORY, "HISTORY.rst")
-    history = open(history_path, "r").read()
+    history = open(history_path).read()
 
     def extend(from_str, line):
         from_str += "\n"
         return history.replace(from_str, from_str + line + "\n")
 
-    history = extend(".. to_doc", """
+    history = extend(".. to_doc", f"""
 ---------------------
-%s.dev0
+{new_version}.dev0
 ---------------------
 
-""" % new_version)
+""")
     open(history_path, "w").write(history)
 
     source_mod_path = os.path.join(PROJECT_DIRECTORY, source_dir, "__init__.py")
-    mod = open(source_mod_path, "r").read()
-    mod = re.sub("__version__ = '[\d\.]+'",
-                 "__version__ = '%s.dev0'" % new_version,
+    mod = open(source_mod_path).read()
+    mod = re.sub(r"__version__ = '[\d\.]+'",
+                 f"__version__ = '{new_version}.dev0'",
                  mod, 1)
     mod = open(source_mod_path, "w").write(mod)
-    shell(["git", "commit", "-m", "Starting work on %s" % new_version,
-           "HISTORY.rst", "%s/__init__.py" % source_dir])
+    shell(["git", "commit", "-m", f"Starting work on {new_version}",
+           "HISTORY.rst", f"{source_dir}/__init__.py"])
 
 
 def shell(cmds, **kwds):